+++ .github/workflows/aks.yaml
@@ -41,10 +41,10 @@ on:
         default: 'hostname/password'
         type: string
       tests_to_run:
-        description: Tests to run (p0_provisioning/p0_importing/support_matrix_provisioning/support_matrix_importing/k8s_chart_support_provisioning/k8s_chart_support_importing)
+        description: Tests to run (p0_provisioning/p0_import/support_matrix_provisioning/support_matrix_import/k8s_chart_support_provisioning/k8s_chart_support_import)
         type: string
         required: true
-        default: p0_provisioning/p0_importing
+        default: p0_provisioning/p0_import
 
 jobs:
   aks-e2e:
@@ -55,7 +55,7 @@ jobs:
       rancher_version: ${{ inputs.rancher_version || '2.9-head' }}
       k3s_version: ${{ inputs.k3s_version || 'v1.28.9+k3s1' }}
       operator_nightly_chart: ${{ inputs.operator_nightly_chart == true || (github.event_name == 'schedule' && true) }}
-      tests_to_run: ${{ inputs.tests_to_run || 'p0_provisioning/p0_importing' }}
+      tests_to_run: ${{ inputs.tests_to_run || 'p0_provisioning/p0_import' }}
       destroy_runner: ${{ inputs.destroy_runner == true || (github.event_name == 'schedule' && true) }}
       runner_template: ${{ inputs.runner_template || 'hosted-prov-e2e-ci-runner-spot-n2-highmem-16-gl-template-v2' }}
       downstream_k8s_minor_version: ${{ inputs.downstream_k8s_minor_version }}
