+++ .github/workflows/main.yaml
@@ -215,8 +215,8 @@ jobs:
         run: |
           make e2e-provisioning-tests
 
-      - name: Importing cluster tests
-        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'p0_importing') }}
+      - name: Import cluster tests
+        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'p0_import') }}
         env:
           RANCHER_HOSTNAME: ${{ env.RANCHER_HOSTNAME }}
           RANCHER_PASSWORD: ${{ env.RANCHER_PASSWORD }}
@@ -235,8 +235,8 @@ jobs:
         run: |
           make e2e-p1-provisioning-tests
 
-      - name: Importing cluster P1 tests
-        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'p1_importing') }}
+      - name: Import cluster P1 tests
+        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'p1_import') }}
         env:
           RANCHER_HOSTNAME: ${{ env.RANCHER_HOSTNAME }}
           RANCHER_PASSWORD: ${{ env.RANCHER_PASSWORD }}
@@ -255,15 +255,15 @@ jobs:
         run: |
           make e2e-support-matrix-provisioning-tests
 
-      - name: Support matrix importing tests
-        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'support_matrix_importing') }}
+      - name: Support matrix import tests
+        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'support_matrix_import') }}
         env:
           RANCHER_HOSTNAME: ${{ env.RANCHER_HOSTNAME }}
           RANCHER_PASSWORD: ${{ env.RANCHER_PASSWORD }}
           CATTLE_TEST_CONFIG: ${{ github.workspace }}/cattle-config-import.yaml
           QASE_RUN_ID: ${{ steps.qase.outputs.qase_run_id }}
         run: |
-          make e2e-support-matrix-importing-tests
+          make e2e-support-matrix-import-tests
 
       - name: K8s Chart Support provisioning tests
         if: ${{ !cancelled() && contains(inputs.tests_to_run, 'k8s_chart_support_provisioning') }}
@@ -277,8 +277,8 @@ jobs:
         run: |
           make e2e-k8s-chart-support-provisioning-tests
 
-      - name: K8s Chart Support importing tests
-        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'k8s_chart_support_importing') }}
+      - name: K8s Chart Support import tests
+        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'k8s_chart_support_import') }}
         env:
           RANCHER_HOSTNAME: ${{ env.RANCHER_HOSTNAME }}
           RANCHER_PASSWORD: ${{ env.RANCHER_PASSWORD }}
@@ -287,7 +287,7 @@ jobs:
           KUBECONFIG: /etc/rancher/k3s/k3s.yaml
           DOWNSTREAM_K8S_MINOR_VERSION: ${{ env.DOWNSTREAM_K8S_MINOR_VERSION }}
         run: |
-          make e2e-k8s-chart-support-importing-tests
+          make e2e-k8s-chart-support-import-tests
 
       - name: Sync provisioning tests
         if: ${{ !cancelled() && contains(inputs.tests_to_run, 'sync_provisioning') }}
@@ -300,8 +300,8 @@ jobs:
         run: |
           make e2e-sync-provisioning-tests
 
-      - name: Sync importing tests
-        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'sync_importing') }}
+      - name: Sync import tests
+        if: ${{ !cancelled() && contains(inputs.tests_to_run, 'sync_import') }}
         env:
           RANCHER_HOSTNAME: ${{ env.RANCHER_HOSTNAME }}
           RANCHER_PASSWORD: ${{ env.RANCHER_PASSWORD }}
@@ -309,7 +309,7 @@ jobs:
           QASE_RUN_ID: ${{ steps.qase.outputs.qase_run_id }}
           DOWNSTREAM_K8S_MINOR_VERSION: ${{ env.DOWNSTREAM_K8S_MINOR_VERSION }}
         run: |
-          make e2e-sync-importing-tests
+          make e2e-sync-import-tests
 
       - name: Finalize Qase Run and publish Results
         env:
